+++ resources/ReleaseNotes.md
@@ -8,5 +8,6 @@
 - Rename `Result` to `PSD`
 - Remove support for `Filters` in charts
 - Allow charts to have `mag.Data` as input
+- Remove implicit conversion methods for `table`, `timetable` and `tabular`
 - Fix issues with setting colors in charts
 - Add tests for area, scatter, stairs and stems plots